requirements.txt
amqp==2.5.0
beautifulsoup4==4.8.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
configparser==3.7.4
geographiclib==1.49
geopy==1.20.0
googlemaps==3.0.2
idna==2.8
kombu==4.6.3
mysql-connector-python==8.0.17
protobuf==3.9.0
pyaes==1.6.1
Pyrogram==0.15.1
PySocks==1.7.0
pytz==2019.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
urllib3==1.25.3
vine==1.3.0
